requirements.txt
pytest==8.3.2
requests==2.32.3
flask==2.3.3
scalekit-sdk-python
PyYAML==6.0.3
